Changed from float to double, make the damage combat action take a damage multipler, changed all combat actions to require an user and a target instead of just a target

+++ Scripts/CombatEntity.cs
@@ -50,7 +50,7 @@ public partial class CombatEntity : Node2D
         Collider.AddChild(ColliderShape);
     }
 
-    public void TakeDamage(float damage)
+    public void TakeDamage(double damage)
     {
         Stats.CurrentHealth -= damage;
         HealthBar.Value = Stats.CurrentHealth / Stats.MaxHealth;
